Basic NPC implementation
Including changes to the skill/fight system, the update and render loop
and the basic Entity implementation. We now have walking and potentialy
killing/defeatable NPCs! But for some reason they won't die, probably
the player is a wimp. Or his attack action is incorrect. But, personaly,
I guess it is the former case.